requirements.txt
# python -m venv ./venv
# source ./venv/bin/activate
# pip install -r ./requirements.txt
# python3 -m ipykernel install --name=venv
advertools==0.5.2
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleads==38.0.0
googleapis-common-protos==1.56.2
gspread
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
numpy==1.21.5
oauthlib==3.2.2
pandas==1.4.2
pandas-gbq
plotly==5.6.0
pycausalimpact==0.1.1
pygsheets
PyPDF2==3.0.1
pytrends==4.9.2
pyusuggest==1.0.5
sentencepiece==0.1.99
transformers==4.31.0
urllib3==1.26.6